+++ packages/benchmark-js/src/Benchmark.ts
@@ -1,6 +1,7 @@
 import { BenchmarkJob } from './BenchmarkJob';
 import { ConsoleLogger, LogKind } from './tools/ConsoleLogger';
 import { BenchmarkJobOptions, TestFn } from './types';
+import { Column, Table } from './View';
 
 export class Benchmark {
     private jobs: BenchmarkJob[] = [];
@@ -24,16 +25,23 @@ export class Benchmark {
 
     public run(): void {
         const logger = ConsoleLogger.default;
-        logger.writeLine(
-            LogKind.Info,
-            `// Found ${this.jobs.length} ${this.jobs.length > 1 ? 'benchmarks' : 'benchmark'}:`,
-        );
+        logger.writeLineInfo(`// Found ${this.jobs.length} ${this.jobs.length > 1 ? 'benchmarks' : 'benchmark'}:`);
         for (const job of this.jobs) {
             logger.writeLine(LogKind.Info, `//   ${job.name}`);
         }
         logger.writeLine();
 
         for (const job of this.jobs) job.run();
-        for (const job of this.jobs) job.writeResult();
+
+        const table = new Table();
+        table.addColumn(new Column('Function', (stats) => stats.name));
+        table.addColumn(new Column('Mean', (stats) => stats.mean));
+        table.addColumn(new Column('StdErr', (stats) => stats.standardError));
+        table.addColumn(new Column('StdDev', (stats) => stats.standardDeviation));
+        table.addColumn(new Column('Median', (stats) => stats.median));
+        table.addColumn(new Column('Min', (stats) => stats.min));
+        table.addColumn(new Column('Max', (stats) => stats.max));
+        for (const job of this.jobs) table.addStats(job.stats);
+        table.draw();
     }
 }